app/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
python-jose
bcrypt
jinja2
python-multipart
kafka-python
